rename: 3d 로고 아이콘 파일 위치 이동 및 파일명 변경 (#14)
